# app/swagger.yaml
openapi: 3.0.0
info:
  title: ODFDS BackEnd API
  description: ODFDS BackEnd API
  version: 1.0.0
servers:
  - url: http://164.92.68.162:3000
    description: Remote Development Server
  - url: http://127.0.0.1:3000
    description: Local Development Server
...
paths:
  /restaurant/emailCode:
    get:
      description: Get an email verification code for creating a restaurant account
      tags:
      - Restaurant
      summary: Get an email code
      parameters:
      - in: query
        name: email
        schema: 
          type: string
        required: true
      responses:
        '200':
          description: Email code sent
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/GeneralResponse'
              example:
                code: 200
                data: Code sent to your email, it will expire in 5 mins
        '400':
          description: Invalid email address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralInvalidResponse'
              examples:
                Invalid email address format:
                  value:
                    code: 400
                    data:
                      field: email
                      value: abc.com
                      message: Invalid email address
                Email address already taken:
                  value:
                    code: 400
                    data:
                      field: email
                      value: steven.panxd@gmail.com
                      message: Email is already taken, please try another email address
                Email code already sent:
                  value:
                    code: 400
                    data:
                      field: email
                      value: steven.panxd@gmail.com
                      message: Code is already sent, please try again after 5 mins
        '500':
          description: Can not send out email, server error
  /restaurant/token:
    post:
      description: Log in an restaurant account and receive a json web token
      tags:
      - Restaurant
      summary: Log in restaurant account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantLoginRequest'
      responses:
        '200':
          description: Login succeed, receive an json web toke
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
              example:
                code: 200
                data: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralInvalidResponse'
              examples:
                Invalid account:
                  value:
                    code: 400
                    data:
                      field: email
                      value: steven.com
                      message: Invalid account
                Account not exist:
                  value:
                    code: 400
                    data:
                      field: email
                      value: steven@gmail.com
                      message: Account does not exist
                Invalid password:
                  value:
                    code: 400
                    data:
                      field: password
                      value: abc123
                      message: Invalid password
                Incorrect password:
                  value:
                    code: 400
                    data:
                      field: password
                      value: A1b2c3.
                      message: Incorrect password
  /restaurant:
    post:
      description: Sign up a new restaurant accout
      tags:
      - Restaurant
      summary: Sign up a new restaurant accout
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantSignUpRequest'
      responses:
        '200':
          description: Sign up succeed, restaurant account data inserted to the database
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/GeneralResponse'
              example:
                code: 200
                data: Restaurant account created successfully
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralInvalidResponse'
              examples:
                Email verification code expired:
                  value: 
                    code: 400
                    data:
                      field: code
                      value: 123456
                      message: Email verification code is expired, please try to request a new one
                Wrong Email verification code:
                  value: 
                    code: 400
                    data:
                      field: code
                      value: 123456
                      message: Email verification code is wrong, please try again
                Email address already taken:
                  value:
                    code: 400
                    data:
                      field: email
                      value: steven.panxd@gmail.com
                      message: Email is already taken, please try another email address
    delete:
      description: Delete a restaurant account (dev use)
      tags:
      - Restaurant
      summary: Delete a restaurant account (dev use)
      parameters:
      - in: query
        name: email
        schema:
          type: string
        required: true
      responses:
        '200':
          description: Succeed
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/GeneralResponse'
              example:
                code: 200
                data: Restaurant account deleted
        '400':
          description: Fail
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/GeneralInvalidResponse'
              examples:
                Invalid email address format:
                  value:
                    code: 400
                    data:
                      field: email
                      value: stevenxpan@yahoo.com
                      message: Restaurant account does not exist
  /restaurant/profile:
    get:
      description: Get a restaurant profile (basic information)
      tags:
      - Restaurant
      summary: Get a restaurant profile
      parameters:
      - in: header
        name: access_token
        schema:
          type: string
          description: restaurant access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        required: true
      responses:
        '200':
          description: return restaurant profile information
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/GetRestaurantProfileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralInvalidResponse'
              example:
                code: 401
                data: Please log in
    patch:
      description: Update restaurant account profile info
      tags:
      - Restaurant
      summary: Update restaurant accout profile
      parameters:
        - in: header
          name: access_token
          schema:
            type: string
            description: restaurant access token
            example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRestaurantProfileRequest'
      responses:
        '200':
          description: Sign up succeed, restaurant account data inserted to the database
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/GeneralResponse'
              example:
                code: 200
                data: Restaurant account created successfully
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralInvalidResponse'
              examples:
                Invalid phone number:
                  value: 
                    code: 400
                    data:
                      field: phone
                      value: ABC
                      message: Invalid phone number
                Duplicated phone number:
                  value: 
                    code: 400
                    data:
                      field: phone
                      value: 4081234567
                      message: Phone number is already taken, please try another phone number
                # Invalid latitude or longitude:
                #   value:
                #     code: 400
                #     data:
                #       field: latitude
                #       value: ABC
                #       message: Invalid latitude
  /restaurant/reset/emailCode:
    get:
      description: Get an email verification code for reset password of a restaurant account
      tags:
      - Restaurant
      summary: Get an email code for reset password for a restaurant account
      parameters:
      - in: query
        name: email
        schema: 
          type: string
        required: true
      responses:
        '200':
          description: Email code sent
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/GeneralResponse'
              example:
                code: 200
                data: Code sent to your email, it will expire in 5 mins
        '400':
          description: Invalid email address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralInvalidResponse'
              examples:
                Invalid email address format:
                  value:
                    code: 400
                    data:
                      field: email
                      value: abc.com
                      message: Invalid email address
                Restaurant account does not exist:
                  value:
                    code: 400
                    data:
                      field: email
                      value: steven.panxd@gmail.com
                      message: Restaurant account does not exist
                Email code already sent:
                  value:
                    code: 400
                    data:
                      field: email
                      value: steven.panxd@gmail.com
                      message: Code is already sent, please try again after 5 mins
        '500':
          description: Can not send out email, server error
  /restaurant/reset/password:
    patch:
      description: Reset restaurant account password
      tags:
      - Restaurant
      summary: Reset restaurant account password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reseted
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/GeneralResponse'
              example:
                code: 200
                data: Succeed
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralInvalidResponse'
              examples:
                Email verification code expired:
                  value: 
                    code: 400
                    data:
                      field: code
                      value: 123456
                      message: Email verification code is expired, please try to request a new one
                Wrong Email verification code:
                  value: 
                    code: 400
                    data:
                      field: code
                      value: 123456
                      message: Email verification code is wrong, please try again
                Restaurant account does not exist:
                  value:
                    code: 400
                    data:
                      field: email
                      value: steven.panxd@gmail.com
                      message: Restaurant account does not exist
  /restaurant/orders:
    get:
      description: Get restaurant history orders
      tags:
      - Restaurant
      summary: Get restaurant history orders
      parameters:
      - in: header
        name: access_token
        schema:
          type: string
          description: restaurant access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        required: true
      - in: query
        name: pageSize
        description: number of items per page
        schema: 
          type: number
        required: true
        example: 10
      - in: query
        name: page
        description: current page number
        required: true
        schema:
          type: number
        example: 1
      responses:
        '200':
          description: return restaurant history orders information
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/GetRestaurantOrdersResponse'
        '400':
          description: invalid parameters
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralInvalidResponse'
              examples:
                  Invalid page number:
                    value:
                      code: 400
                      data:
                        field: page
                        value: abc
                        message: Invalid page number
                  Invalid page size:
                    value:
                      code: 400
                      data:
                        field: pageSize
                        value: abc
                        message: Invalid page size
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralInvalidResponse'
              example:
                code: 401
                data: Please log in
  /restaurant/order/estimate:
    post:
      description: get estimated distance and duration between restaurant and customer
      tags:
      - Restaurant
      summary: get estimated distance and duration
      parameters:
        - in: header
          name: access_token
          schema:
            type: string
            description: restaurant access token
            example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantEstimateRequest'
      responses:
        '200':
          description: Succeed, received the estimated distance (in meters) and duration (in seconds)
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/GeneralResponse'
              example:
                code: 200
                data:
                    estimatedDistanceInMeters: 1555
                    estimatedDurationInSeconds: 234
                    estimatedPriceInDollars: 22.86
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralInvalidResponse'
              examples:
                No street:
                  value: 
                    code: 400
                    data:
                      field: street
                      value: null
                      message: Please input street
                No city:
                  value: 
                    code: 400
                    data:
                      field: city
                      value: null
                      message: Please input city
                No state:
                  value: 
                    code: 400
                    data:
                      field: state
                      value: null
                      message: Please input state
                No zipCode:
                  value: 
                    code: 400
                    data:
                      field: zipCode
                      value: null
                      message: Please input zip code
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralInvalidResponse'
              example:
                code: 401
                data: Please log in
  /restaurant/order:
    post:
      description: create a delivery order
      tags:
      - Restaurant
      summary: create a delivery order
      parameters:
        - in: header
          name: access_token
          schema:
            type: string
            description: restaurant access token
            example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantCreateOrderRequest'
      responses:
        '200':
          description: Succeed, return order id
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/GeneralResponse'
              example:
                code: 200
                data:
                    message: Order created
                    orderId: 1
...
components:
  schemas:
    GeneralResponse:
      title: General valid response body
      type: object
      properties:
        code:
          type: number
          description: HTTP status code
          example: 200
        data:
          type: string
          description: response data or succeed message
          example: succeed
    GeneralInvalidResponse:
      title: General invalid parameter response body
      type: object
      properties:
        code:
          type: number
          description: HTTP status code
          example: 400
        data:
          type: object
          properties:
            field:
              type: string
              description: error field
            value:
              type: string
              description: backend received error value
            message:
              type: string
              description: error message
          example:
            field: email
            value: abc.com
            message: Invalid email address
    GetRestaurantProfileResponse:
      title: Restaurant profile response body
      type: object
      properties:
        code:
          type: number
          description: HTTP status code
          example: 200
        data:
          type: object
          properties:
            id:
              type: number
              description: restaurant account id
              example: 1
            email:
              type: string
              description: email address, unique
              example: steven.panxd@gmail.com
            phone:
              type: string
              description: phone number, unique
              example: 4081234567
            name:
              type: string
              description: restaurant name
              example: My Restaurant
            street:
              type: string
              description: restaurant street
              example: 1 Washington Square
            city:
              type: string
              description: restaurant city
              example: San Jose
            state:
              type: string
              description: restaurant state
              example: California
            zipCode:
              type: string
              description: restaurant zipCode
              example: 95192
            verification:
              type: string
              description: restaurant verification status
              example: NOTVERIFIED
            # latitude:
            #   type: number
            #   description: restaurant latitude
            #   example: 37.33482353097262
            # longitude:
            #   type: number
            #   description: restaurant longitude
            #   example: -121.88420310757643
...
    RestaurantLoginRequest:
      title: Restaurant account log in request body
      type: object
      properties:
        email:
          type: string
          description: email address
          example: steven.panxd@gmail.com
        password:
          type: string
          description: password, much be length >= 6, at least 1 uppercase letter, 1 lowercase letter, and 1 symbol
          example: A1b2c.
    RestaurantSignUpRequest:
      title: Restaurant account sign up request body
      type: object
      properties:
        email:
          type: string
          description: email address, unique
          example: steven.panxd@gmail.com
        code:
          type: string
          description: email verification code
          example: 123456
        password:
          type: string
          description: password, much be length >= 6, at least 1 uppercase letter, 1 lowercase letter, and 1 symbol
          example: A1b2c.
        phone:
          type: string
          description: phone number, unique
          example: 4081234567
        name:
          type: string
          description: restaurant name
          example: My Restaurant
        street:
          type: string
          description: restaurant street
          example: 1 Washington Square
        city:
          type: string
          description: restaurant city
          example: San Jose
        state:
          type: string
          description: restaurant state
          example: California
        zipCode:
          type: string
          description: restaurant zipCode
          example: 95192
        # latitude:
        #   type: number
        #   description: restaurant latitude
        #   example: 37.33482353097262
        # longitude:
        #   type: number
        #   description: restaurant longitude
        #   example: -121.88420310757643
      required:
        - email
        - code
        - password
        - phone
        - name
        - street
        - state
        - zipCode
        # - latitude
        # - longitude
...
    ResetPasswordRequest:
      title: Restaurant/Driver account reset password request body
      type: object
      properties:
        email:
          type: string
          description: email address, unique
          example: steven.panxd@gmail.com
        code:
          type: string
          description: email verification code
          example: 123456
        password:
          type: string
          description: password, much be length >= 6, at least 1 uppercase letter, 1 lowercase letter, and 1 symbol
          example: A1b2c.
      required:
      - email
      - code
      - password
...